FSL Match and Try is a admin endpoint
