+++ go/permissions/servicepermissions.go
@@ -62,6 +62,7 @@ func GetServicePermissions() *ServicePermissions {
 					"/metalstack.admin.v2.FilesystemService/Create",
 					"/metalstack.admin.v2.FilesystemService/Update",
 					"/metalstack.admin.v2.FilesystemService/Delete",
+					"/metalstack.admin.v2.FilesystemService/Match",
 					"/metalstack.admin.v2.ImageService/Create",
 					"/metalstack.admin.v2.ImageService/Update",
 					"/metalstack.admin.v2.ImageService/Delete",
@@ -119,6 +120,7 @@ func GetServicePermissions() *ServicePermissions {
 					"/metalstack.admin.v2.AuditService/List",
 					"/metalstack.admin.v2.ComponentService/Get",
 					"/metalstack.admin.v2.ComponentService/List",
+					"/metalstack.admin.v2.FilesystemService/Match",
 					"/metalstack.admin.v2.ImageService/Usage",
 					"/metalstack.admin.v2.IPService/List",
 					"/metalstack.admin.v2.MachineService/Get",
@@ -277,6 +279,7 @@ func GetServicePermissions() *ServicePermissions {
 			"/metalstack.admin.v2.ComponentService/List":                     true,
 			"/metalstack.admin.v2.FilesystemService/Create":                  true,
 			"/metalstack.admin.v2.FilesystemService/Delete":                  true,
+			"/metalstack.admin.v2.FilesystemService/Match":                   true,
 			"/metalstack.admin.v2.FilesystemService/Update":                  true,
 			"/metalstack.admin.v2.IPService/List":                            true,
 			"/metalstack.admin.v2.ImageService/Create":                       true,
@@ -333,7 +336,6 @@ func GetServicePermissions() *ServicePermissions {
 			"/metalstack.api.v2.AuditService/List":                           true,
 			"/metalstack.api.v2.FilesystemService/Get":                       true,
 			"/metalstack.api.v2.FilesystemService/List":                      true,
-			"/metalstack.api.v2.FilesystemService/Match":                     true,
 			"/metalstack.api.v2.HealthService/Get":                           true,
 			"/metalstack.api.v2.IPService/Create":                            true,
 			"/metalstack.api.v2.IPService/Delete":                            true,
@@ -425,7 +427,6 @@ func GetServicePermissions() *ServicePermissions {
 			Self: map[string]bool{
 				"/metalstack.api.v2.FilesystemService/Get":          true,
 				"/metalstack.api.v2.FilesystemService/List":         true,
-				"/metalstack.api.v2.FilesystemService/Match":        true,
 				"/metalstack.api.v2.ImageService/Get":               true,
 				"/metalstack.api.v2.ImageService/Latest":            true,
 				"/metalstack.api.v2.ImageService/List":              true,
@@ -458,6 +459,7 @@ func GetServicePermissions() *ServicePermissions {
 				"/metalstack.admin.v2.ComponentService/List":             true,
 				"/metalstack.admin.v2.FilesystemService/Create":          true,
 				"/metalstack.admin.v2.FilesystemService/Delete":          true,
+				"/metalstack.admin.v2.FilesystemService/Match":           true,
 				"/metalstack.admin.v2.FilesystemService/Update":          true,
 				"/metalstack.admin.v2.IPService/List":                    true,
 				"/metalstack.admin.v2.ImageService/Create":               true,
@@ -583,6 +585,7 @@ func GetServicePermissions() *ServicePermissions {
 			"/metalstack.admin.v2.ComponentService/List":             false,
 			"/metalstack.admin.v2.FilesystemService/Create":          true,
 			"/metalstack.admin.v2.FilesystemService/Delete":          true,
+			"/metalstack.admin.v2.FilesystemService/Match":           false,
 			"/metalstack.admin.v2.FilesystemService/Update":          true,
 			"/metalstack.admin.v2.IPService/List":                    false,
 			"/metalstack.admin.v2.ImageService/Create":               true,
@@ -639,7 +642,6 @@ func GetServicePermissions() *ServicePermissions {
 			"/metalstack.api.v2.AuditService/List":                   false,
 			"/metalstack.api.v2.FilesystemService/Get":               false,
 			"/metalstack.api.v2.FilesystemService/List":              false,
-			"/metalstack.api.v2.FilesystemService/Match":             false,
 			"/metalstack.api.v2.HealthService/Get":                   false,
 			"/metalstack.api.v2.IPService/Create":                    true,
 			"/metalstack.api.v2.IPService/Delete":                    true,
