+++ js/metalstack/api/v2/filesystem_pb.js
@@ -8,7 +8,7 @@ import { file_metalstack_api_v2_predefined_rules } from "./predefined_rules_pb";
 /**
  * Describes the file metalstack/api/v2/filesystem.proto.
  */
-export const file_metalstack_api_v2_filesystem = /*@__PURE__*/ fileDesc("CiJtZXRhbHN0YWNrL2FwaS92Mi9maWxlc3lzdGVtLnByb3RvEhFtZXRhbHN0YWNrLmFwaS52MiI2ChtGaWxlc3lzdGVtU2VydmljZUdldFJlcXVlc3QSFwoCaWQYASABKAlCC7pICHIGwLOusQIBIkMKHEZpbGVzeXN0ZW1TZXJ2aWNlTGlzdFJlcXVlc3QSHAoCaWQYASABKAlCC7pICHIGwLOusQIBSACIAQFCBQoDX2lkIl4KHEZpbGVzeXN0ZW1TZXJ2aWNlR2V0UmVzcG9uc2USPgoRZmlsZXN5c3RlbV9sYXlvdXQYASABKAsyIy5tZXRhbHN0YWNrLmFwaS52Mi5GaWxlc3lzdGVtTGF5b3V0ImAKHUZpbGVzeXN0ZW1TZXJ2aWNlTGlzdFJlc3BvbnNlEj8KEmZpbGVzeXN0ZW1fbGF5b3V0cxgBIAMoCzIjLm1ldGFsc3RhY2suYXBpLnYyLkZpbGVzeXN0ZW1MYXlvdXQiywEKHUZpbGVzeXN0ZW1TZXJ2aWNlTWF0Y2hSZXF1ZXN0Ej4KDnNpemVfYW5kX2ltYWdlGAEgASgLMiQubWV0YWxzdGFjay5hcGkudjIuTWF0Y2hJbWFnZUFuZFNpemVIABJaChxtYWNoaW5lX2FuZF9maWxlc3lzdGVtbGF5b3V0GAIgASgLMjIubWV0YWxzdGFjay5hcGkudjIuTWF0Y2hNYWNoaW5lQW5kRmlsZXN5c3RlbUxheW91dEgAQg4KBW1hdGNoEgW6SAIIASJKChFNYXRjaEltYWdlQW5kU2l6ZRIZCgRzaXplGAEgASgJQgu6SAhyBsCzrrECARIaCgVpbWFnZRgCIAEoCUILukgIcgbAs66xAgEiZAofTWF0Y2hNYWNoaW5lQW5kRmlsZXN5c3RlbUxheW91dBIZCgdtYWNoaW5lGAEgASgJQgi6SAVyA7ABARImChFmaWxlc3lzdGVtX2xheW91dBgCIAEoCUILukgIcgbAs66xAgEiYAoeRmlsZXN5c3RlbVNlcnZpY2VNYXRjaFJlc3BvbnNlEj4KEWZpbGVzeXN0ZW1fbGF5b3V0GAEgASgLMiMubWV0YWxzdGFjay5hcGkudjIuRmlsZXN5c3RlbUxheW91dCLsAwoQRmlsZXN5c3RlbUxheW91dBIXCgJpZBgBIAEoCUILukgIcgbAs66xAgESJQoEbWV0YRgCIAEoCzIXLm1ldGFsc3RhY2suYXBpLnYyLk1ldGESHgoEbmFtZRgDIAEoCUILukgIcgbAs66xAgFIAIgBARIlCgtkZXNjcmlwdGlvbhgEIAEoCUILukgIcgbIs66xAgFIAYgBARIyCgtmaWxlc3lzdGVtcxgFIAMoCzIdLm1ldGFsc3RhY2suYXBpLnYyLkZpbGVzeXN0ZW0SJgoFZGlza3MYBiADKAsyFy5tZXRhbHN0YWNrLmFwaS52Mi5EaXNrEiUKBHJhaWQYByADKAsyFy5tZXRhbHN0YWNrLmFwaS52Mi5SYWlkEjUKDXZvbHVtZV9ncm91cHMYCCADKAsyHi5tZXRhbHN0YWNrLmFwaS52Mi5Wb2x1bWVHcm91cBI5Cg9sb2dpY2FsX3ZvbHVtZXMYCSADKAsyIC5tZXRhbHN0YWNrLmFwaS52Mi5Mb2dpY2FsVm9sdW1lEkMKC2NvbnN0cmFpbnRzGAogASgLMi4ubWV0YWxzdGFjay5hcGkudjIuRmlsZXN5c3RlbUxheW91dENvbnN0cmFpbnRzQgcKBV9uYW1lQg4KDF9kZXNjcmlwdGlvbiKnAQobRmlsZXN5c3RlbUxheW91dENvbnN0cmFpbnRzEg0KBXNpemVzGAEgAygJEkoKBmltYWdlcxgCIAMoCzI6Lm1ldGFsc3RhY2suYXBpLnYyLkZpbGVzeXN0ZW1MYXlvdXRDb25zdHJhaW50cy5JbWFnZXNFbnRyeRotCgtJbWFnZXNFbnRyeRILCgNrZXkYASABKAkSDQoFdmFsdWUYAiABKAk6AjgBIuwCCgpGaWxlc3lzdGVtEhsKBmRldmljZRgBIAEoCUILukgIcgbAs66xAgESNgoGZm9ybWF0GAIgASgOMhkubWV0YWxzdGFjay5hcGkudjIuRm9ybWF0Qgu6SAjIAQGCAQIQARIeCgRuYW1lGAMgASgJQgu6SAhyBsCzrrECAUgAiAEBEiUKC2Rlc2NyaXB0aW9uGAQgASgJQgu6SAhyBsizrrECAUgBiAEBEh0KBHBhdGgYBSABKAlCCrpIB3IFEAEYgCBIAogBARIeCgVsYWJlbBgGIAEoCUIKukgHcgUQAhiAAUgDiAEBEioKDW1vdW50X29wdGlvbnMYByADKAlCE7pIEJIBDRAgGAEiB3IFEAEYgAESKwoOY3JlYXRlX29wdGlvbnMYCCADKAlCE7pIEJIBDRAgGAEiB3IFEAEYgAFCBwoFX25hbWVCDgoMX2Rlc2NyaXB0aW9uQgcKBV9wYXRoQggKBl9sYWJlbCJYCgREaXNrEhoKBmRldmljZRgBIAEoCUIKukgHcgUQAhiAARI0CgpwYXJ0aXRpb25zGAIgAygLMiAubWV0YWxzdGFjay5hcGkudjIuRGlza1BhcnRpdGlvbiKiAQoEUmFpZBISCgphcnJheV9uYW1lGAEgASgJEg8KB2RldmljZXMYAiADKAkSOAoFbGV2ZWwYAyABKA4yHC5tZXRhbHN0YWNrLmFwaS52Mi5SYWlkTGV2ZWxCC7pICMgBAYIBAhABEisKDmNyZWF0ZV9vcHRpb25zGAQgAygJQhO6SBCSAQ0QIBgBIgdyBRABGIABEg4KBnNwYXJlcxgFIAEoBSKYAQoNRGlza1BhcnRpdGlvbhIOCgZudW1iZXIYASABKA0SEgoFbGFiZWwYAiABKAlIAIgBARIMCgRzaXplGAMgASgEEj4KCGdwdF90eXBlGAQgASgOMhoubWV0YWxzdGFjay5hcGkudjIuR1BUVHlwZUILukgIyAEBggECEAFIAYgBAUIICgZfbGFiZWxCCwoJX2dwdF90eXBlIkcKC1ZvbHVtZUdyb3VwEhkKBG5hbWUYASABKAlCC7pICHIGwLOusQIBEg8KB2RldmljZXMYAiADKAkSDAoEdGFncxgDIAMoCSKSAQoNTG9naWNhbFZvbHVtZRIZCgRuYW1lGAEgASgJQgu6SAhyBsCzrrECARIgCgx2b2x1bWVfZ3JvdXAYAiABKAlCCrpIB3IFEAIYgAESDAoEc2l6ZRgDIAEoBBI2Cghsdm1fdHlwZRgEIAEoDjIaLm1ldGFsc3RhY2suYXBpLnYyLkxWTVR5cGVCCLpIBYIBAhABKoYBCgdMVk1UeXBlEhgKFExWTV9UWVBFX1VOU1BFQ0lGSUVEEAASHwoPTFZNX1RZUEVfTElORUFSEAEaCoKyGQZsaW5lYXISIQoQTFZNX1RZUEVfU1RSSVBFRBACGguCshkHc3RyaXBlZBIdCg5MVk1fVFlQRV9SQUlEMRADGgmCshkFcmFpZDEqxAEKBkZvcm1hdBIWChJGT1JNQVRfVU5TUEVDSUZJRUQQABIZCgtGT1JNQVRfVkZBVBABGgiCshkEdmZhdBIZCgtGT1JNQVRfRVhUMxACGgiCshkEZXh0MxIZCgtGT1JNQVRfRVhUNBADGgiCshkEZXh0NBIZCgtGT1JNQVRfU1dBUBAEGgiCshkEc3dhcBIbCgxGT1JNQVRfVE1QRlMQBRoJgrIZBXRtcGZzEhkKC0ZPUk1BVF9OT05FEAYaCIKyGQRub25lKqMBCgdHUFRUeXBlEhgKFEdQVF9UWVBFX1VOU1BFQ0lGSUVEEAASGwoNR1BUX1RZUEVfQk9PVBABGgiCshkEZWYwMBIcCg5HUFRfVFlQRV9MSU5VWBACGgiCshkEODMwMBIhChNHUFRfVFlQRV9MSU5VWF9SQUlEEAMaCIKyGQRmZDAwEiAKEkdQVF9UWVBFX0xJTlVYX0xWTRAEGgiCshkEOGUwMCpLCglSYWlkTGV2ZWwSGgoWUkFJRF9MRVZFTF9VTlNQRUNJRklFRBAAEhAKDFJBSURfTEVWRUxfMBABEhAKDFJBSURfTEVWRUxfMRACMvICChFGaWxlc3lzdGVtU2VydmljZRJwCgNHZXQSLi5tZXRhbHN0YWNrLmFwaS52Mi5GaWxlc3lzdGVtU2VydmljZUdldFJlcXVlc3QaLy5tZXRhbHN0YWNrLmFwaS52Mi5GaWxlc3lzdGVtU2VydmljZUdldFJlc3BvbnNlIgjY8xgC4PMYAhJzCgRMaXN0Ei8ubWV0YWxzdGFjay5hcGkudjIuRmlsZXN5c3RlbVNlcnZpY2VMaXN0UmVxdWVzdBowLm1ldGFsc3RhY2suYXBpLnYyLkZpbGVzeXN0ZW1TZXJ2aWNlTGlzdFJlc3BvbnNlIgjY8xgC4PMYAhJ2CgVNYXRjaBIwLm1ldGFsc3RhY2suYXBpLnYyLkZpbGVzeXN0ZW1TZXJ2aWNlTWF0Y2hSZXF1ZXN0GjEubWV0YWxzdGFjay5hcGkudjIuRmlsZXN5c3RlbVNlcnZpY2VNYXRjaFJlc3BvbnNlIgjY8xgC4PMYAkLFAQoVY29tLm1ldGFsc3RhY2suYXBpLnYyQg9GaWxlc3lzdGVtUHJvdG9QAVo1Z2l0aHViLmNvbS9tZXRhbC1zdGFjay9hcGkvZ28vbWV0YWxzdGFjay9hcGkvdjI7YXBpdjKiAgNNQViqAhFNZXRhbHN0YWNrLkFwaS5WMsoCEU1ldGFsc3RhY2tcQXBpXFYy4gIdTWV0YWxzdGFja1xBcGlcVjJcR1BCTWV0YWRhdGHqAhNNZXRhbHN0YWNrOjpBcGk6OlYyYgZwcm90bzM", [file_buf_validate_validate, file_metalstack_api_v2_common, file_metalstack_api_v2_predefined_rules]);
+export const file_metalstack_api_v2_filesystem = /*@__PURE__*/ fileDesc("CiJtZXRhbHN0YWNrL2FwaS92Mi9maWxlc3lzdGVtLnByb3RvEhFtZXRhbHN0YWNrLmFwaS52MiI2ChtGaWxlc3lzdGVtU2VydmljZUdldFJlcXVlc3QSFwoCaWQYASABKAlCC7pICHIGwLOusQIBIkMKHEZpbGVzeXN0ZW1TZXJ2aWNlTGlzdFJlcXVlc3QSHAoCaWQYASABKAlCC7pICHIGwLOusQIBSACIAQFCBQoDX2lkIl4KHEZpbGVzeXN0ZW1TZXJ2aWNlR2V0UmVzcG9uc2USPgoRZmlsZXN5c3RlbV9sYXlvdXQYASABKAsyIy5tZXRhbHN0YWNrLmFwaS52Mi5GaWxlc3lzdGVtTGF5b3V0ImAKHUZpbGVzeXN0ZW1TZXJ2aWNlTGlzdFJlc3BvbnNlEj8KEmZpbGVzeXN0ZW1fbGF5b3V0cxgBIAMoCzIjLm1ldGFsc3RhY2suYXBpLnYyLkZpbGVzeXN0ZW1MYXlvdXQi7AMKEEZpbGVzeXN0ZW1MYXlvdXQSFwoCaWQYASABKAlCC7pICHIGwLOusQIBEiUKBG1ldGEYAiABKAsyFy5tZXRhbHN0YWNrLmFwaS52Mi5NZXRhEh4KBG5hbWUYAyABKAlCC7pICHIGwLOusQIBSACIAQESJQoLZGVzY3JpcHRpb24YBCABKAlCC7pICHIGyLOusQIBSAGIAQESMgoLZmlsZXN5c3RlbXMYBSADKAsyHS5tZXRhbHN0YWNrLmFwaS52Mi5GaWxlc3lzdGVtEiYKBWRpc2tzGAYgAygLMhcubWV0YWxzdGFjay5hcGkudjIuRGlzaxIlCgRyYWlkGAcgAygLMhcubWV0YWxzdGFjay5hcGkudjIuUmFpZBI1Cg12b2x1bWVfZ3JvdXBzGAggAygLMh4ubWV0YWxzdGFjay5hcGkudjIuVm9sdW1lR3JvdXASOQoPbG9naWNhbF92b2x1bWVzGAkgAygLMiAubWV0YWxzdGFjay5hcGkudjIuTG9naWNhbFZvbHVtZRJDCgtjb25zdHJhaW50cxgKIAEoCzIuLm1ldGFsc3RhY2suYXBpLnYyLkZpbGVzeXN0ZW1MYXlvdXRDb25zdHJhaW50c0IHCgVfbmFtZUIOCgxfZGVzY3JpcHRpb24ipwEKG0ZpbGVzeXN0ZW1MYXlvdXRDb25zdHJhaW50cxINCgVzaXplcxgBIAMoCRJKCgZpbWFnZXMYAiADKAsyOi5tZXRhbHN0YWNrLmFwaS52Mi5GaWxlc3lzdGVtTGF5b3V0Q29uc3RyYWludHMuSW1hZ2VzRW50cnkaLQoLSW1hZ2VzRW50cnkSCwoDa2V5GAEgASgJEg0KBXZhbHVlGAIgASgJOgI4ASLsAgoKRmlsZXN5c3RlbRIbCgZkZXZpY2UYASABKAlCC7pICHIGwLOusQIBEjYKBmZvcm1hdBgCIAEoDjIZLm1ldGFsc3RhY2suYXBpLnYyLkZvcm1hdEILukgIyAEBggECEAESHgoEbmFtZRgDIAEoCUILukgIcgbAs66xAgFIAIgBARIlCgtkZXNjcmlwdGlvbhgEIAEoCUILukgIcgbIs66xAgFIAYgBARIdCgRwYXRoGAUgASgJQgq6SAdyBRABGIAgSAKIAQESHgoFbGFiZWwYBiABKAlCCrpIB3IFEAIYgAFIA4gBARIqCg1tb3VudF9vcHRpb25zGAcgAygJQhO6SBCSAQ0QIBgBIgdyBRABGIABEisKDmNyZWF0ZV9vcHRpb25zGAggAygJQhO6SBCSAQ0QIBgBIgdyBRABGIABQgcKBV9uYW1lQg4KDF9kZXNjcmlwdGlvbkIHCgVfcGF0aEIICgZfbGFiZWwiWAoERGlzaxIaCgZkZXZpY2UYASABKAlCCrpIB3IFEAIYgAESNAoKcGFydGl0aW9ucxgCIAMoCzIgLm1ldGFsc3RhY2suYXBpLnYyLkRpc2tQYXJ0aXRpb24iogEKBFJhaWQSEgoKYXJyYXlfbmFtZRgBIAEoCRIPCgdkZXZpY2VzGAIgAygJEjgKBWxldmVsGAMgASgOMhwubWV0YWxzdGFjay5hcGkudjIuUmFpZExldmVsQgu6SAjIAQGCAQIQARIrCg5jcmVhdGVfb3B0aW9ucxgEIAMoCUITukgQkgENECAYASIHcgUQARiAARIOCgZzcGFyZXMYBSABKAUimAEKDURpc2tQYXJ0aXRpb24SDgoGbnVtYmVyGAEgASgNEhIKBWxhYmVsGAIgASgJSACIAQESDAoEc2l6ZRgDIAEoBBI+CghncHRfdHlwZRgEIAEoDjIaLm1ldGFsc3RhY2suYXBpLnYyLkdQVFR5cGVCC7pICMgBAYIBAhABSAGIAQFCCAoGX2xhYmVsQgsKCV9ncHRfdHlwZSJHCgtWb2x1bWVHcm91cBIZCgRuYW1lGAEgASgJQgu6SAhyBsCzrrECARIPCgdkZXZpY2VzGAIgAygJEgwKBHRhZ3MYAyADKAkikgEKDUxvZ2ljYWxWb2x1bWUSGQoEbmFtZRgBIAEoCUILukgIcgbAs66xAgESIAoMdm9sdW1lX2dyb3VwGAIgASgJQgq6SAdyBRACGIABEgwKBHNpemUYAyABKAQSNgoIbHZtX3R5cGUYBCABKA4yGi5tZXRhbHN0YWNrLmFwaS52Mi5MVk1UeXBlQgi6SAWCAQIQASqGAQoHTFZNVHlwZRIYChRMVk1fVFlQRV9VTlNQRUNJRklFRBAAEh8KD0xWTV9UWVBFX0xJTkVBUhABGgqCshkGbGluZWFyEiEKEExWTV9UWVBFX1NUUklQRUQQAhoLgrIZB3N0cmlwZWQSHQoOTFZNX1RZUEVfUkFJRDEQAxoJgrIZBXJhaWQxKsQBCgZGb3JtYXQSFgoSRk9STUFUX1VOU1BFQ0lGSUVEEAASGQoLRk9STUFUX1ZGQVQQARoIgrIZBHZmYXQSGQoLRk9STUFUX0VYVDMQAhoIgrIZBGV4dDMSGQoLRk9STUFUX0VYVDQQAxoIgrIZBGV4dDQSGQoLRk9STUFUX1NXQVAQBBoIgrIZBHN3YXASGwoMRk9STUFUX1RNUEZTEAUaCYKyGQV0bXBmcxIZCgtGT1JNQVRfTk9ORRAGGgiCshkEbm9uZSqjAQoHR1BUVHlwZRIYChRHUFRfVFlQRV9VTlNQRUNJRklFRBAAEhsKDUdQVF9UWVBFX0JPT1QQARoIgrIZBGVmMDASHAoOR1BUX1RZUEVfTElOVVgQAhoIgrIZBDgzMDASIQoTR1BUX1RZUEVfTElOVVhfUkFJRBADGgiCshkEZmQwMBIgChJHUFRfVFlQRV9MSU5VWF9MVk0QBBoIgrIZBDhlMDAqSwoJUmFpZExldmVsEhoKFlJBSURfTEVWRUxfVU5TUEVDSUZJRUQQABIQCgxSQUlEX0xFVkVMXzAQARIQCgxSQUlEX0xFVkVMXzEQAjL6AQoRRmlsZXN5c3RlbVNlcnZpY2UScAoDR2V0Ei4ubWV0YWxzdGFjay5hcGkudjIuRmlsZXN5c3RlbVNlcnZpY2VHZXRSZXF1ZXN0Gi8ubWV0YWxzdGFjay5hcGkudjIuRmlsZXN5c3RlbVNlcnZpY2VHZXRSZXNwb25zZSII2PMYAuDzGAIScwoETGlzdBIvLm1ldGFsc3RhY2suYXBpLnYyLkZpbGVzeXN0ZW1TZXJ2aWNlTGlzdFJlcXVlc3QaMC5tZXRhbHN0YWNrLmFwaS52Mi5GaWxlc3lzdGVtU2VydmljZUxpc3RSZXNwb25zZSII2PMYAuDzGAJCxQEKFWNvbS5tZXRhbHN0YWNrLmFwaS52MkIPRmlsZXN5c3RlbVByb3RvUAFaNWdpdGh1Yi5jb20vbWV0YWwtc3RhY2svYXBpL2dvL21ldGFsc3RhY2svYXBpL3YyO2FwaXYyogIDTUFYqgIRTWV0YWxzdGFjay5BcGkuVjLKAhFNZXRhbHN0YWNrXEFwaVxWMuICHU1ldGFsc3RhY2tcQXBpXFYyXEdQQk1ldGFkYXRh6gITTWV0YWxzdGFjazo6QXBpOjpWMmIGcHJvdG8z", [file_buf_validate_validate, file_metalstack_api_v2_common, file_metalstack_api_v2_predefined_rules]);
 /**
  * Describes the message metalstack.api.v2.FilesystemServiceGetRequest.
  * Use `create(FilesystemServiceGetRequestSchema)` to create a new message.
@@ -30,65 +30,45 @@ export const FilesystemServiceGetResponseSchema = /*@__PURE__*/ messageDesc(file
  */
 export const FilesystemServiceListResponseSchema = /*@__PURE__*/ messageDesc(file_metalstack_api_v2_filesystem, 3);
 /**
- * Describes the message metalstack.api.v2.FilesystemServiceMatchRequest.
- * Use `create(FilesystemServiceMatchRequestSchema)` to create a new message.
- */
-export const FilesystemServiceMatchRequestSchema = /*@__PURE__*/ messageDesc(file_metalstack_api_v2_filesystem, 4);
-/**
- * Describes the message metalstack.api.v2.MatchImageAndSize.
- * Use `create(MatchImageAndSizeSchema)` to create a new message.
- */
-export const MatchImageAndSizeSchema = /*@__PURE__*/ messageDesc(file_metalstack_api_v2_filesystem, 5);
-/**
- * Describes the message metalstack.api.v2.MatchMachineAndFilesystemLayout.
- * Use `create(MatchMachineAndFilesystemLayoutSchema)` to create a new message.
- */
-export const MatchMachineAndFilesystemLayoutSchema = /*@__PURE__*/ messageDesc(file_metalstack_api_v2_filesystem, 6);
-/**
- * Describes the message metalstack.api.v2.FilesystemServiceMatchResponse.
- * Use `create(FilesystemServiceMatchResponseSchema)` to create a new message.
- */
-export const FilesystemServiceMatchResponseSchema = /*@__PURE__*/ messageDesc(file_metalstack_api_v2_filesystem, 7);
-/**
  * Describes the message metalstack.api.v2.FilesystemLayout.
  * Use `create(FilesystemLayoutSchema)` to create a new message.
  */
-export const FilesystemLayoutSchema = /*@__PURE__*/ messageDesc(file_metalstack_api_v2_filesystem, 8);
+export const FilesystemLayoutSchema = /*@__PURE__*/ messageDesc(file_metalstack_api_v2_filesystem, 4);
 /**
  * Describes the message metalstack.api.v2.FilesystemLayoutConstraints.
  * Use `create(FilesystemLayoutConstraintsSchema)` to create a new message.
  */
-export const FilesystemLayoutConstraintsSchema = /*@__PURE__*/ messageDesc(file_metalstack_api_v2_filesystem, 9);
+export const FilesystemLayoutConstraintsSchema = /*@__PURE__*/ messageDesc(file_metalstack_api_v2_filesystem, 5);
 /**
  * Describes the message metalstack.api.v2.Filesystem.
  * Use `create(FilesystemSchema)` to create a new message.
  */
-export const FilesystemSchema = /*@__PURE__*/ messageDesc(file_metalstack_api_v2_filesystem, 10);
+export const FilesystemSchema = /*@__PURE__*/ messageDesc(file_metalstack_api_v2_filesystem, 6);
 /**
  * Describes the message metalstack.api.v2.Disk.
  * Use `create(DiskSchema)` to create a new message.
  */
-export const DiskSchema = /*@__PURE__*/ messageDesc(file_metalstack_api_v2_filesystem, 11);
+export const DiskSchema = /*@__PURE__*/ messageDesc(file_metalstack_api_v2_filesystem, 7);
 /**
  * Describes the message metalstack.api.v2.Raid.
  * Use `create(RaidSchema)` to create a new message.
  */
-export const RaidSchema = /*@__PURE__*/ messageDesc(file_metalstack_api_v2_filesystem, 12);
+export const RaidSchema = /*@__PURE__*/ messageDesc(file_metalstack_api_v2_filesystem, 8);
 /**
  * Describes the message metalstack.api.v2.DiskPartition.
  * Use `create(DiskPartitionSchema)` to create a new message.
  */
-export const DiskPartitionSchema = /*@__PURE__*/ messageDesc(file_metalstack_api_v2_filesystem, 13);
+export const DiskPartitionSchema = /*@__PURE__*/ messageDesc(file_metalstack_api_v2_filesystem, 9);
 /**
  * Describes the message metalstack.api.v2.VolumeGroup.
  * Use `create(VolumeGroupSchema)` to create a new message.
  */
-export const VolumeGroupSchema = /*@__PURE__*/ messageDesc(file_metalstack_api_v2_filesystem, 14);
+export const VolumeGroupSchema = /*@__PURE__*/ messageDesc(file_metalstack_api_v2_filesystem, 10);
 /**
  * Describes the message metalstack.api.v2.LogicalVolume.
  * Use `create(LogicalVolumeSchema)` to create a new message.
  */
-export const LogicalVolumeSchema = /*@__PURE__*/ messageDesc(file_metalstack_api_v2_filesystem, 15);
+export const LogicalVolumeSchema = /*@__PURE__*/ messageDesc(file_metalstack_api_v2_filesystem, 11);
 /**
  * LVMType
  *
