+++ python/metalstack/api/v2/filesystem_connect.py
@@ -22,9 +22,6 @@ class FilesystemService(Protocol):
     async def list(self, request: metalstack_dot_api_dot_v2_dot_filesystem__pb2.FilesystemServiceListRequest, ctx: RequestContext) -> metalstack_dot_api_dot_v2_dot_filesystem__pb2.FilesystemServiceListResponse:
         raise ConnectError(Code.UNIMPLEMENTED, "Not implemented")
 
-    async def match(self, request: metalstack_dot_api_dot_v2_dot_filesystem__pb2.FilesystemServiceMatchRequest, ctx: RequestContext) -> metalstack_dot_api_dot_v2_dot_filesystem__pb2.FilesystemServiceMatchResponse:
-        raise ConnectError(Code.UNIMPLEMENTED, "Not implemented")
-
 
 class FilesystemServiceASGIApplication(ConnectASGIApplication[FilesystemService]):
     def __init__(self, service: FilesystemService | AsyncGenerator[FilesystemService], *, interceptors: Iterable[Interceptor]=(), read_max_bytes: int | None = None) -> None:
@@ -50,16 +47,6 @@ class FilesystemServiceASGIApplication(ConnectASGIApplication[FilesystemService]
                         idempotency_level=IdempotencyLevel.UNKNOWN,
                     ),
                     function=svc.list,
-                ),
-                "/metalstack.api.v2.FilesystemService/Match": Endpoint.unary(
-                    method=MethodInfo(
-                        name="Match",
-                        service_name="metalstack.api.v2.FilesystemService",
-                        input=metalstack_dot_api_dot_v2_dot_filesystem__pb2.FilesystemServiceMatchRequest,
-                        output=metalstack_dot_api_dot_v2_dot_filesystem__pb2.FilesystemServiceMatchResponse,
-                        idempotency_level=IdempotencyLevel.UNKNOWN,
-                    ),
-                    function=svc.match,
                 ),
             },
             interceptors=interceptors,
@@ -113,33 +100,11 @@ class FilesystemServiceClient(ConnectClient):
             timeout_ms=timeout_ms,
         )
 
-    async def match(
-        self,
-        request: metalstack_dot_api_dot_v2_dot_filesystem__pb2.FilesystemServiceMatchRequest,
-        *,
-        headers: Headers | Mapping[str, str] | None = None,
-        timeout_ms: int | None = None,
-    ) -> metalstack_dot_api_dot_v2_dot_filesystem__pb2.FilesystemServiceMatchResponse:
-        return await self.execute_unary(
-            request=request,
-            method=MethodInfo(
-                name="Match",
-                service_name="metalstack.api.v2.FilesystemService",
-                input=metalstack_dot_api_dot_v2_dot_filesystem__pb2.FilesystemServiceMatchRequest,
-                output=metalstack_dot_api_dot_v2_dot_filesystem__pb2.FilesystemServiceMatchResponse,
-                idempotency_level=IdempotencyLevel.UNKNOWN,
-            ),
-            headers=headers,
-            timeout_ms=timeout_ms,
-        )
-
 
 class FilesystemServiceSync(Protocol):
     def get(self, request: metalstack_dot_api_dot_v2_dot_filesystem__pb2.FilesystemServiceGetRequest, ctx: RequestContext) -> metalstack_dot_api_dot_v2_dot_filesystem__pb2.FilesystemServiceGetResponse:
         raise ConnectError(Code.UNIMPLEMENTED, "Not implemented")
     def list(self, request: metalstack_dot_api_dot_v2_dot_filesystem__pb2.FilesystemServiceListRequest, ctx: RequestContext) -> metalstack_dot_api_dot_v2_dot_filesystem__pb2.FilesystemServiceListResponse:
-        raise ConnectError(Code.UNIMPLEMENTED, "Not implemented")
-    def match(self, request: metalstack_dot_api_dot_v2_dot_filesystem__pb2.FilesystemServiceMatchRequest, ctx: RequestContext) -> metalstack_dot_api_dot_v2_dot_filesystem__pb2.FilesystemServiceMatchResponse:
         raise ConnectError(Code.UNIMPLEMENTED, "Not implemented")
 
 
@@ -166,16 +131,6 @@ class FilesystemServiceWSGIApplication(ConnectWSGIApplication):
                         idempotency_level=IdempotencyLevel.UNKNOWN,
                     ),
                     function=service.list,
-                ),
-                "/metalstack.api.v2.FilesystemService/Match": EndpointSync.unary(
-                    method=MethodInfo(
-                        name="Match",
-                        service_name="metalstack.api.v2.FilesystemService",
-                        input=metalstack_dot_api_dot_v2_dot_filesystem__pb2.FilesystemServiceMatchRequest,
-                        output=metalstack_dot_api_dot_v2_dot_filesystem__pb2.FilesystemServiceMatchResponse,
-                        idempotency_level=IdempotencyLevel.UNKNOWN,
-                    ),
-                    function=service.match,
                 ),
             },
             interceptors=interceptors,
@@ -223,26 +178,6 @@ class FilesystemServiceClientSync(ConnectClientSync):
                 service_name="metalstack.api.v2.FilesystemService",
                 input=metalstack_dot_api_dot_v2_dot_filesystem__pb2.FilesystemServiceListRequest,
                 output=metalstack_dot_api_dot_v2_dot_filesystem__pb2.FilesystemServiceListResponse,
-                idempotency_level=IdempotencyLevel.UNKNOWN,
-            ),
-            headers=headers,
-            timeout_ms=timeout_ms,
-        )
-
-    def match(
-        self,
-        request: metalstack_dot_api_dot_v2_dot_filesystem__pb2.FilesystemServiceMatchRequest,
-        *,
-        headers: Headers | Mapping[str, str] | None = None,
-        timeout_ms: int | None = None,
-    ) -> metalstack_dot_api_dot_v2_dot_filesystem__pb2.FilesystemServiceMatchResponse:
-        return self.execute_unary(
-            request=request,
-            method=MethodInfo(
-                name="Match",
-                service_name="metalstack.api.v2.FilesystemService",
-                input=metalstack_dot_api_dot_v2_dot_filesystem__pb2.FilesystemServiceMatchRequest,
-                output=metalstack_dot_api_dot_v2_dot_filesystem__pb2.FilesystemServiceMatchResponse,
                 idempotency_level=IdempotencyLevel.UNKNOWN,
             ),
             headers=headers,
